Version 6 - working
Finds optimal path using A* algorithm and Euclidean norm of difference
between current and goal as heuristic function. Prints the path used at the
end of the script as an an array and prints a matrix showing nodes
in the path as 1s, and nodes not in the path as 0s. Also now prints a matrix
showing which nodes have been visited each loop iteration.